Revert "Try GUI test again", nope, not testing GUI in CI
This reverts commit b2477d052ca27298819e91e45580b341eb32b310.
