+++ .github/workflows/tests.yml
@@ -40,7 +40,5 @@ jobs:
           python -m pip install --upgrade pip
           python -m pip install -r requirements_CI.txt
           pip install tox tox-gh-actions
-          sudo rm /usr/lib/x86_64-linux-gnu/libEGL.so
-          sudo ln /usr/lib/x86_64-linux-gnu/libEGL.so.1 /usr/lib/x86_64-linux-gnu/libEGL.so
       - name: Test with tox
         run: tox
