+++ gui.py
@@ -6,6 +6,6 @@ Run with ./gui.py. If that doesn't work, run however you normally run a Python s
 
 # Imports from NeuroRuler/GUI/__init__.py
 # This is exactly what the user can access after installing via pip
-from NeuroRuler import gui
+from NeuroRuler.GUI import gui
 
 gui()
